Updated React command to extend AbstractCommand and implemented Service
