+++ src/modules/commands/React/index.ts
@@ -1,16 +1,8 @@
 import * as Discord from "discord.js";
-import fs from "fs";
-import * as FileHelper from "../../io/FileHelper";
-import Path from "path";
+import ReactService from "./Service/ReactService";
+import AbstractCommand from "../../../models/AbstractCommand";
 
-import CommandInterface  from "../../../models/CommandInterface";
-import settings from "../../../settings.json";
-
-import constants from "../../constants";
-
-const reactImgFolder = "./assets/img/reaction/";
-
-class React implements CommandInterface {
+class React extends AbstractCommand {
     public name = "react";
     public description = "Allows you to use reaction images.";
     public usage = ["add <name>", "remove <name>", "<name>"];
@@ -21,137 +13,18 @@ class React implements CommandInterface {
      * @param {Discord.Message} message The Discord message which called the command.
      */
     public execute(message: Discord.Message): void {
-        const commandStr = message.content.split(constants.COMMAND_PREFIX)[1];
-        const parameters = commandStr.split(" ");
+        const commandParameters = super.getCommandParameters();
 
-        // Switch based on sub-command
-        switch (parameters[1]) {
+        switch (commandParameters[1]) {
             case "add":
-                this.addReaction(parameters[2], message);
+                ReactService.addReaction(commandParameters[2], message);
                 break;
             case "remove":
-                this.removeReaction(parameters[2], message);
+                ReactService.removeReaction(commandParameters[2], message);
                 break;
             default:
-                this.sendReaction(parameters[1], message);
+                ReactService.sendReaction(commandParameters[1], message);
                 break;
-        }
-    }
-
-    /**
-     * Handler for "react [name]".
-     *
-     * Sends a message with the reaction defined by name
-     * to the channel where the command was called.
-     *
-     * @param {String} name The name of the reaction.
-     * @param {Discord.Message} message The Discord message which called the command.
-     */
-    private sendReaction(name: string, message: Discord.Message): void {
-        const author = message.author;
-        const channel = message.channel;
-
-        // Retrieves all filenames in reaction image folder and finds the file specified.
-        const availableReactions = FileHelper.getReactionImageList();
-        let reactionFilename = availableReactions.find((file) => file.split(".")[0] === name);
-
-        // Assure filename is safe and not PathLike
-        if (reactionFilename) {
-            reactionFilename = Path.basename(reactionFilename);
-        }
-
-        // Sends the file if it was found, otherwise sends error message.
-        if (reactionFilename !== undefined) {
-            channel.send({ files: [reactImgFolder + reactionFilename] });
-        } else {
-            channel.send(`${author}.\n\n:x: The reaction \`${name}\` does not exist. :x:`);
-        }
-    }
-
-    /**
-     * Handler for "react add [name]".
-     *
-     * Saves the attached file as a new reaction with the given name. Afterwards sends
-     * back success/error message to channel where the command was called.
-     *
-     * This function requires the user calling it to be in the settings.json admin_list.
-     *
-     * @param {String} name The name of the emote to be added.
-     * @param {Discord.Message} message: The message which called the command.
-     */
-    private addReaction(name: string, message: Discord.Message): void {
-        const admins = settings.admin_list;
-        const author = message.author;
-        const attachments = message.attachments;
-        const channel = message.channel;
-
-        // Retrieves all filenames in reaction image folder and finds the file specified.
-        const availableReactions = FileHelper.getReactionImageList();
-        let reactionFilename = availableReactions.find((file) => file.split(".")[0] === name);
-
-        // Check if user is in the admin list in settings.json.
-        if (!admins.includes(author.id)) {
-            channel.send(`:x: Sorry ${author}, this command can only be run by an administrator. :x:`);
-            return;
-        }
-
-        // Assure filename is safe and not PathLike
-        if (reactionFilename) {
-            reactionFilename = Path.basename(reactionFilename);
-        }
-
-        // Creates the file if it was not found, otherwise sends error message.
-        if (reactionFilename === undefined) {
-            FileHelper.saveDiscordMessageAttachments(name, reactImgFolder, attachments);
-
-            channel.send(
-                `:white_check_mark: Created reaction \`${name}\`! You can now use it with \`${constants.COMMAND_PREFIX}react ${name}\`! :white_check_mark:`,
-            );
-        } else {
-            channel.send(
-                `${author}.\n\n:x: The reaction \`${name}\` already exists. Please use another name or remove the current reaction first. :x:`,
-            );
-        }
-    }
-
-    /**
-     * Handler for "react remove [name]".
-     *
-     * Removes the reaction with the given name from the bot. Afterwards sends
-     * back success/error message to channel where the command was called.
-     *
-     * This function requires the user calling it to be in the settings.json admin_list.
-     *
-     * @param {String} name The name of the emote to be added.
-     * @param {Discord.Message} message: The message which called the command.
-     */
-    private removeReaction(name: string, message: Discord.Message): void {
-        const admins = settings.admin_list;
-        const author = message.author;
-        const channel = message.channel;
-
-        // Retrieves all filenames in reaction image folder and finds the file specified.
-        const availableReactions = FileHelper.getReactionImageList();
-        let reactionFilename = availableReactions.find((file) => file.split(".")[0] === name);
-
-        // Check if user is in the admin list in settings.json.
-        if (!admins.includes(author.id)) {
-            channel.send(`:x: Sorry ${author}, this command can only be run by an administrator. :x:`);
-            return;
-        }
-
-        // Assure filename is safe and not PathLike
-        if (reactionFilename) {
-            reactionFilename = Path.basename(reactionFilename);
-        }
-
-        // Creates the file if it was not found, otherwise sends error message.
-        if (reactionFilename !== undefined) {
-            fs.unlinkSync(reactImgFolder + reactionFilename);
-
-            channel.send(`:white_check_mark: Removed reaction \`${name}\`! :white_check_mark:`);
-        } else {
-            channel.send(`${author}.\n\n:x: The reaction \`${name}\` does not exist. :x:`);
         }
     }
 }
